Changed the GetTotalPrice to return a decimal rather than an int. Added the code to calculate the totalPrice in the concrete class but failing the GetTotalPrice_Returns_Correct_Price_For_Scanned_Products test

+++ src/BrightHR.Checkout/ICheckout.cs
@@ -7,6 +7,6 @@ namespace BrightHR.Checkout
         public IList<string> ScannedProducts { get; }
 
         void Scan(string item);
-        int GetTotalPrice();
+        decimal GetTotalPrice();
     }
 }
